requirements.txt
psycopg2-binary
redis
flask